+++ api-service/src/main/java/com/sobok/apiservice/api/service/socialLogin/GoogleLoginService.java
@@ -5,6 +5,7 @@ import com.sobok.apiservice.api.client.GoogleFeignClient;
 import com.sobok.apiservice.api.client.GoogleTokenFeignClient;
 import com.sobok.apiservice.api.dto.google.*;
 import com.sobok.apiservice.api.dto.kakao.OauthResDto;
+import com.sobok.apiservice.api.dto.social.SocialUserDto;
 import com.sobok.apiservice.api.entity.Oauth;
 import com.sobok.apiservice.api.repository.OauthRepository;
 import com.sobok.apiservice.common.dto.ApiResponse;
@@ -35,8 +36,7 @@ public class GoogleLoginService {
 
     private final GoogleFeignClient googleFeignClient;
     private final GoogleTokenFeignClient googleTokenFeignClient;
-    private final OauthRepository oauthRepository;
-    private final AuthFeignClient authFeignClient;
+    private final SocialLoginService socialLoginService;
 
     //백엔드에서 구글 로그인 화면(authorization URL)을 생성해서 프론트에 전달해주는 메서드
     //사용자가 구글 로그인을 시도할 때 구글 인증 페이지로 리다이렉션할 수 있는 URL을 만들어 주는 역할
@@ -71,9 +71,12 @@ public class GoogleLoginService {
         // 인가 코드에서 얻은 액세스토큰으로 사용자 정보 받기
         GoogleDetailResDto googleDetailResDto = loginGoogle(code);
         log.info("받은 googleDetailResDto 사용자 정보: {}", googleDetailResDto);
-
+        SocialUserDto socialDto = SocialUserDto.builder()
+                .socialId(googleDetailResDto.getSub())
+                .provider("GOOGLE")
+                .build();
         // 회원가입 or 로그인 처리
-        OauthResDto oauthResDto = findOrCreateKakaoUser(googleDetailResDto);  //authId와 닉네임
+        OauthResDto oauthResDto = socialLoginService.findOrCreateKakaoUser(socialDto);  //authId와 닉네임
         log.info("oauthResDto: {}", oauthResDto);
 
         return GoogleCallResDto.builder()
@@ -83,59 +86,6 @@ public class GoogleLoginService {
                 .email(googleDetailResDto.getEmail())
                 .picture(googleDetailResDto.getPicture())
                 .isNew(oauthResDto.isNew())
-                .build();
-    }
-
-    @Transactional
-    public OauthResDto findOrCreateKakaoUser(GoogleDetailResDto dto) {
-        // 기존 구글 로그인 사용자
-        // 구글 ID로 기존 사용자 찾기
-        Optional<Oauth> oauth = oauthRepository.findBySocialProviderAndSocialId("GOOGLE", dto.getSub());
-
-        log.info("dto.getSub(): {} oauth가 있다면: {}", dto.getSub(), oauth);
-
-        // 기존 사용자 존재
-        if (oauth.isPresent()) {
-            log.info("기존에 구글 소셜 로그인한 유저입니다.");
-            Oauth foundUser = oauth.get();
-            OauthResDto oauthResDto;
-            try {
-                oauthResDto = authFeignClient.authIdById(foundUser.getId());
-            } catch (FeignException.NotFound e) {
-                log.info("oauth는 생성하였지만 회원가입을 마치지 못한 유저입니다.");
-                return OauthResDto.builder()
-                        .oauthId(foundUser.getId())
-                        .socialId(foundUser.getSocialId())
-                        .isNew(true)
-                        .build();
-            } catch (FeignException e) {
-                log.error("Feign 호출 중 예외 발생", e);
-                throw new CustomException("OAuth 정보 조회 중 오류 발생", HttpStatus.INTERNAL_SERVER_ERROR);
-            }
-            log.info("oauthResDto: {}", oauthResDto);
-            return OauthResDto.builder()
-                    .oauthId(foundUser.getId())
-                    .socialId(foundUser.getSocialId())
-                    .authId(oauthResDto.getAuthId())
-                    .isNew(false)
-                    .build();
-        }
-
-        // 처음 구글 로그인 한 사람 -> 새 사용자 생성. oauth
-        log.info("구글 로그인으로 처음 방문한 신규 유저입니다. 회원가입 진행해야 됨");
-
-        Oauth build = Oauth.builder()
-                .socialProvider("GOOGLE")
-                .socialId(dto.getSub())
-                .build();
-
-        Oauth saved = oauthRepository.save(build);
-
-        log.info("saved: {}", saved);
-
-        return OauthResDto.builder()
-                .oauthId(saved.getId())
-                .isNew(true)
                 .build();
     }
