+++ api-service/src/main/java/com/sobok/apiservice/api/service/socialLogin/KakaoLoginService.java
@@ -2,6 +2,7 @@ package com.sobok.apiservice.api.service.socialLogin;
 
 import com.sobok.apiservice.api.client.AuthFeignClient;
 import com.sobok.apiservice.api.dto.kakao.*;
+import com.sobok.apiservice.api.dto.social.SocialUserDto;
 import com.sobok.apiservice.api.entity.Oauth;
 import com.sobok.apiservice.api.repository.OauthRepository;
 import com.sobok.apiservice.common.exception.CustomException;
@@ -29,9 +30,7 @@ public class KakaoLoginService {
     @Value("${oauth2.kakao.redirect-uri}")
     private String kakaoRedirectUri;
 
-    private final AuthFeignClient authFeignClient;
-    private final OauthRepository oauthRepository;
-
+    private final SocialLoginService socialLoginService;
 
     // 인가 코드로 카카오 액세스 토큰 받기
     public String getKakaoAccessToken(String code) {
@@ -95,85 +94,17 @@ public class KakaoLoginService {
 
     }
 
-    @Transactional
-    public OauthResDto findOrCreateKakaoUser(KakaoUserResDto dto) {
-        // 기존 카카오 로그인 사용자
-        // 카카오 ID로 기존 사용자 찾기
-        Optional<Oauth> oauth = oauthRepository.findBySocialProviderAndSocialId("KAKAO", dto.getId().toString());
-
-        log.info("dto.getId(): {} oauth가 있을까요 없을까요: {}", dto.getId(), oauth);
-
-        // 기존 사용자 존재
-        if (oauth.isPresent()) {
-            log.info("기존에 카카오 소셜 로그인한 유저입니다.");
-            Oauth foundUser = oauth.get();
-            OauthResDto oauthResDto;
-            try {
-                oauthResDto = authFeignClient.authIdById(foundUser.getId());
-            } catch (FeignException.NotFound e) {
-                log.info("oauth는 생성하였지만 회원가입을 마치지 못한 유저입니다.");
-                return OauthResDto.builder()
-                        .oauthId(foundUser.getId())
-                        .socialId(foundUser.getSocialId())
-                        .isNew(true)
-                        .build();
-            } catch (FeignException e) {
-                log.error("Feign 호출 중 예외 발생", e);
-                throw new CustomException("OAuth 정보 조회 중 오류 발생", HttpStatus.INTERNAL_SERVER_ERROR);
-            }
-            log.info("oauthResDto: {}", oauthResDto);
-            return OauthResDto.builder()
-                    .oauthId(foundUser.getId())
-                    .socialId(foundUser.getSocialId())
-                    .authId(oauthResDto.getAuthId())
-                    .isNew(false)
-                    .build();
-        }
-
-        // 기존 계정(email)으로 가입한 유저인지 확인
-        // 구현 안함
-
-        // 처음 카카오 로그인 한 사람 -> 새 사용자 생성. oauth
-        log.info("카카오 로그인으로 처음 방문한 신규 유저입니다. 회원가입 진행해야 됨");
-
-        Oauth build = Oauth.builder()
-                .socialProvider("KAKAO")
-                .socialId(dto.getId().toString())
-                .build();
-
-        Oauth saved = oauthRepository.save(build);
-
-        log.info("saved: {}", saved);
-
-        return OauthResDto.builder()
-                .oauthId(saved.getId())
-                .isNew(true)
-                .build();
-    }
-
-    public AuthLoginResDto socialLoginToken(Long id) {  //authId
-        return authFeignClient.socialToken(id);
-    }
-
-    public OauthResDto findOauth(Long oauthId) {
-        Oauth oauth = oauthRepository.findById(oauthId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 Oauth ID의 사용자를 찾을 수 없습니다."));
-
-        log.info("oauth: {}", oauth);
-
-        return OauthResDto.builder()
-                .oauthId(oauth.getId())
-                .socialId(oauth.getSocialId())
-                .build();
-    }
-
     public KakaoCallResDto kakaoCallback(String code) {
         // 인가코드로 액세스토큰 받기
         String kakaoAccessToken = getKakaoAccessToken(code);
         // 액세스토큰으로 사용자 정보 받기
         KakaoUserResDto kakaoUserDto = getKakaoUserInfo(kakaoAccessToken);
+        SocialUserDto socialDto = SocialUserDto.builder()
+                .socialId(kakaoUserDto.getId().toString())
+                .provider("KAKAO")
+                .build();
         // 회원가입 or 로그인 처리
-        OauthResDto oauthResDto = findOrCreateKakaoUser(kakaoUserDto);  //authId와 닉네임
+        OauthResDto oauthResDto = socialLoginService.findOrCreateKakaoUser(socialDto);  //authId와 닉네임
 
         log.info("oauthResDto: {}", oauthResDto);
